Add workaround for DTR _catalog API strangeness
DTR used to return the repository information from the _catalog API, but
now it has started not to and returns an empty list instead. When we
detect an empty repository response, we will try to use the DTR
repositories API instead as a fallback. If that fails, then we accept
that maybe there legitimately are no repositories in the registry.